photo_colorizer/requirements.txt
# PyTorch with CUDA 12.1 support (RTX 3070 / Ampere)
# Install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch
torchvision

# Colorization models
colorizers            # Zhang et al. ECCV 2016 (reliable fallback)

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Color palette extraction
scikit-learn>=1.3.0

# Web UI
gradio>=4.0.0

# DDColor dependencies (optional – install manually if available)
# pip install ddcolor
# If ddcolor is unavailable, the app automatically falls back to colorizers.
